debugging

+++ src/utils/cloudinary.js
@@ -14,7 +14,7 @@ const uploadOnCloudinary = async (localFilePath) => {
         fs.unlinkSync(localFilePath)
         return response;
     } catch (error) {
-        fs.unlinkSync(localFilePath)
+        if (fs.existsSync(localFilePath)) fs.unlinkSync(localFilePath)
         return null;
     }
 }
